src/backend/ai/tts/service.py

### TTS Service Implementation:

```python
# src/backend/ai/tts/service.py
import os
import logging
from typing import Optional, List, Dict, Any
from fastapi import FastAPI, HTTPException, BackgroundTasks
from pydantic import BaseModel
import json
import numpy as np
from datetime import datetime
import base64

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

app = FastAPI(title="TTS Service")

# Models for requests and responses
class SynthesisRequest(BaseModel):
    text: str
    language: Optional[str] = "en-US"
    voice: Optional[str] = "female-1"
    sample_rate: Optional[int] = 44100
    speaking_rate: Optional[float] = 1.0
    pitch: Optional[float] = 0.0
    output_path: Optional[str] = None

class SynthesisResponse(BaseModel):
    audio_path: Optional[str] = None
    audio_base64: Optional[str] = None
    sample_rate: int
    duration: float
    processing_time: float
    format: str = "wav"

# Riva TTS client setup
try:
    import riva.client
    riva_client = riva.client.SpeechSynthesisClient(
        uri="localhost:50051",
        use_ssl=False,
        ssl_cert=None
    )
    logger.info("Riva TTS client initialized successfully")
except Exception as e:
    logger.error(f"Failed to initialize Riva TTS client: {str(e)}")
    riva_client = None

@app.get("/health")
def health_check():
    """Health check endpoint"""
    return {
        "status": "healthy",
        "service": "tts-service",
        "riva_available": riva_client is not None
    }

@app.get("/voices")
def list_voices():
    """List available TTS voices"""
    if not riva_client:
        raise HTTPException(status_code=503, detail="TTS service not available")
    
    try:
        voices = riva_client.list_voices()
        return {
            "voices": [
                {
                    "name": voice.name,
                    "language_codes": voice.language_codes,
                    "voice_type": str(voice.voice_type)
                } for voice in voices.voices
            ]
        }
    except Exception as e:
        logger.error(f"Error listing voices: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/synthesize", response_model=SynthesisResponse)
async def synthesize_speech(request: SynthesisRequest):
    """Synthesize speech using Riva TTS"""
    if not riva_client:
        raise HTTPException(status_code=503, detail="TTS service not available")
    
    try:
        start_time = datetime.now()
        
        # Configure TTS parameters
        config = riva.client.SynthesisConfig()
        config.language_code = request.language
        config.voice_name = request.voice
        config.sample_rate_hz = request.sample_rate
        config.speaking_rate = request.speaking_rate
        config.pitch = request.pitch
        
        # Perform synthesis
        response = riva_client.synthesize(
            request.text,
            config=config
        )
        
        # Process results
        audio_samples = np.frombuffer(response.audio, dtype=np.int16)
        duration = len(audio_samples) / request.sample_rate
        
        # Determine output path
        output_path = request.output_path
        if not output_path:
            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
            output_path = f"/app/tts_data/speech_{timestamp}.wav"
        
        # Save audio data
        import wave
        with wave.open(output_path, 'wb') as wf:
            wf.setnchannels(1)
            wf.setsampwidth(2)  # 16-bit audio
            wf.setframerate(request.sample_rate)
            wf.writeframes(response.audio)
        
        # Encode audio to base64 for web delivery
        audio_base64 = base64.b64encode(response.audio).decode('utf-8')
        
        processing_time = (datetime.now() - start_time).total_seconds()
        
        return SynthesisResponse(
            audio_path=output_path,
            audio_base64=audio_base64,
            sample_rate=request.sample_rate,
            duration=duration,
            processing_time=processing_time
        )
    
    except Exception as e:
        logger.error(f"Error during speech synthesis: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    import uvicorn
    port = int(os.environ.get("PORT", 8002))
    uvicorn.run("tts.service:app", host="0.0.0.0", port=port, reload=False)
```