requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1732732445543/work
annotated-types @ file:///croot/annotated-types_1709542908624/work
anyio==4.9.0
archspec @ file:///croot/archspec_1709217642129/work
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
boltons @ file:///croot/boltons_1737061692168/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1738623731865/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
conda @ file:///croot/conda_1738168370816/work
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1736524596117/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1739298990519/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1737733694612/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
contourpy==1.3.1
cryptography @ file:///croot/cryptography_1732129644106/work
cycler==0.12.1
datasets==3.5.0
decorator==5.2.1
dill==0.3.8
distro @ file:///croot/distro_1714488253808/work
executing==2.2.0
fastapi==0.115.11
filelock==3.18.0
fonttools==4.56.0
frozendict @ file:///croot/frozendict_1713194832637/work
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==8.6.1
iniconfig==2.1.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.21
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langsmith==0.3.19
lazy_loader==0.4
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
librosa==0.11.0
litellm==1.65.0
llvmlite==0.44.0
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1698846045803/work
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1698845653285/work
menuinst @ file:///croot/menuinst_1738943416351/work
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvidia-riva-client==2.19.0
openai==1.70.0
orjson==3.10.16
packaging @ file:///croot/packaging_1734472117206/work
pandarallel==1.6.5
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///croot/pluggy_1733169602837/work
pooch==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///croot/pydantic_1734736067156/work
pydantic_core @ file:///croot/pydantic-core_1734726052986/work
Pygments @ file:///work/perseverance-python-buildout/croot/pygments_1698846270603/work
pyparsing==3.2.3
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
requests-toolbelt==1.0.0
rich @ file:///croot/rich_1732638981168/work
-e git+https://github.com/lm-sys/RouteLLM.git@0b64fdafe049e596a3f5657c219329f24af24198#egg=routellm
rpds-py==0.24.0
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setproctitle==1.3.5
setuptools==70.0.0
sglang==0.4.4.post3
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
tqdm @ file:///croot/tqdm_1738943501192/work
traitlets==5.14.3
transformers==4.50.0
triton==3.2.0
truststore @ file:///croot/truststore_1736550121485/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1737133630106/work
uvicorn==0.34.0
wcwidth==0.2.13
wheel==0.45.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard @ file:///croot/zstandard_1731356346222/work
